chore: move tests to corresponding file
